// client/one/app/core/entities/types/campaign/campaign-budgets-overview.ts
export interface CampaignBudgetsOverview {
    campaignSpend: string;
    mediaSpend: string;
    baseMediaSpend: string;
    dataSpend: string;
    baseDataSpend: string;
    licenseFee: string;
    serviceFee: string;
    margin: string;
    availableBudgetsSum: string;
    unallocatedCredit: string;
}